add storage-service, notification-service, user-service

+++ server/domain/src/main/kotlin/com/sudoo/domain/base/Pagination.kt
@@ -2,5 +2,5 @@ package com.sudoo.domain.base
 
 data class Pagination(
     val offset: Int = 0,
-    val total: Int = 0,
+    val total: Long = 0,
 )
